Committing oldest snapshot saved as first commit
This is the oldest snapshot I have of my game engine prior to using
version control.
I am making this the first commit, and will proceed by committing each
following snapshot one at a time until caught up to date with all
snapshots.
After that, I will proceed to put the entire history of using version
control that occurred after the last snapshot was saved before I made
the switch.
This will recreate the entire history of my development of the game
engine.
Note: This game engine is referred to as "DASGameEngine", which is a
straight continuation of "GameEngine". The rename does not reflect any
changes made.