2015.02.17 - Base Game Engine defined
At this point in development, I was able to label my project as a "game
engine in progress" for the first time. Albeit extremely barebones, the
project could render an object, and had room for continuous engine
growth.
By industry standards, the project was no where near what anyone would
even think of labeling as a game engine, but the work had passed the
tutorial stage and was now onto the engine development stage.
This is the snapshot at which point tutorial elements (taken from
multiple tutorials) were starting to be replaced with my own work. Some
of my own work existed prior to this snapshot, but not enough to be able
to claim this as my own work.