2015.02.19-11:05 Window Walls
Playing around with assets, working towards making it so that they can
be shaped as needed, as well as rendered as needed. Currently Models are
becoming parent classes for future classes to inherit from. New shaders
are being made, while older shaders are being repurposed to stay
relevant with the changes.
The ModelAssetBgMesh class has been removed as it is slow and bloated.