2015.02.24 Quaternions
Switched from completing object hierarchy to working on quaternions, as
I realised that they are integral to continuing development. The object
hierarchy has been partially created, and is at a point where it does
not impede with development, and has a clear defined path of how to be
continued when returned to in the future.