2015.03.02 Pre-github First Commit
This was the final snapshot taken before I switched to github. After
this, I did restart my github repository once, and so this will not be
the last commit of snapshots. Rather one last set of snapshots will be
committed, and then the current (not new) repository will be committed
to this new repository.