2015.07.13 Prior to Making New Repo
This is my planned final commit prior to making the complete switch over
to the new DASGameEngine repository. If any more commits occur after
this before a commit message acknowledges being the first intended
commit in the new repo, those commits would have to be due to the
discovery of more up-to-date files that aren't in my snapshot folder.